Refactors, separate ModSetting
