+++ ZHMModSDK/Src/ModLoader.cpp
@@ -604,47 +604,3 @@ void ModLoader::WriteModConfigurationSettingsToIni(mINI::INIMap<std::string>& p_
         p_Ini.set(s_Setting.first, s_Setting.second.Get());
     }
 }
-
-
-ModSetting::~ModSetting()
-{
-}
-
-bool ModSetting::Read(ZString& p_Out)
-{
-    p_Out = ZString(Get());
-    return true;
-}
-
-bool ModSetting::ReadBool(bool& p_Out)
-{
-    p_Out = AsBool();
-    return true;
-}
-
-bool ModSetting::ReadInt(int32& p_Out)
-{
-    auto s_Result = AsInt<int32>();
-    if (!s_Result)
-        return false;
-    p_Out = *s_Result;
-    return true;
-}
-
-bool ModSetting::ReadUnsignedInt(uint32& p_Out)
-{
-    auto s_Result = AsInt<uint32>();
-    if (!s_Result)
-        return false;
-    p_Out = *s_Result;
-    return true;
-}
-
-bool ModSetting::ReadDouble(float64& p_Out)
-{
-    auto s_Result = AsFloat<float64>();
-    if (!s_Result)
-        return false;
-    p_Out = *s_Result;
-    return true;
-}
